ldgorm: updated the path for gorm log
